Address my own comments in PR 173. All tests pass

+++ data_structures/histogram.py
@@ -54,14 +54,14 @@ class Histogram:
 
         # Note: labels can be any type like string or list
         if self.is_classification:
-            self.left = np.zeros((self.num_bins, len(classes)), dtype=np.int32)
-            self.right = np.zeros((self.num_bins, len(classes)), dtype=np.int32)
+            self.left = np.zeros((self.num_bins, len(classes)), dtype=np.int64)
+            self.right = np.zeros((self.num_bins, len(classes)), dtype=np.int64)
         else:
             # self.left_pile[i], self.right_pile[i] stores [number of previous samples, mean of previous samples,
             # variance of previous samples] that are on the left and right of ith bin
-            self.left_pile = np.zeros((self.num_bins, 3))
-            self.right_pile = np.zeros((self.num_bins, 3))
-            self.curr_pile = np.zeros(3)
+            self.left_pile = np.zeros((self.num_bins, 3), dtype=np.int64)
+            self.right_pile = np.zeros((self.num_bins, 3), dtype=np.int64)
+            self.curr_pile = np.zeros(3, dtype=np.int64)
 
     def get_bin(self, fvals: np.ndarray, bin_edges: np.ndarray) -> np.ndarray:
         """
@@ -100,7 +100,7 @@ class Histogram:
             self.left_pile[bin_idcs, :] = np.zeros(3, dtype=np.int64)
             self.right_pile[bin_idcs, :] = np.zeros(3, dtype=np.int64)
             if is_curr_empty:
-                self.curr_pile = np.zeros(3)
+                self.curr_pile = np.zeros(3, dtype=np.int64)
 
     def add(self, X: np.ndarray, Y: np.ndarray):
         """
