+++ utils/utils.py
@@ -46,7 +46,9 @@ def count_occurrence(class_: np.ndarray, labels: np.ndarray) -> int:
     return len(np.where(labels == class_)[0])
 
 
-def class_to_idx(classes: np.ndarray,) -> dict:
+def class_to_idx(
+    classes: np.ndarray,
+) -> dict:
     """
     Helpful function for generating dictionary that maps class names to class index
     Helper function for function for generating dictionary that maps class names to class index
@@ -66,7 +68,7 @@ def counts_of_labels(class_dict: dict, labels: np.ndarray) -> np.ndarray:
     :return: array of counts of each class label, indexed by class index
     """
     classes = np.unique(labels)
-    counts = np.zeros(len(class_dict))
+    counts = np.zeros(len(class_dict), dtype=np.int64)
     for class_ in classes:
         class_idx = class_dict[class_]
         counts[class_idx] = count_occurrence(class_, labels)
@@ -241,6 +243,4 @@ def empty_histograms(histograms: List[Histogram], arms: Tuple[np.ndarray, np.nda
         histogram = histograms[f]
         # Since we don't obviate bins in arms, even though they are not candidates
         # Todo: change this if we obviate bins later
-        histogram.empty_samples(
-            range(histogram.num_bins)
-        )
+        histogram.empty_samples(range(histogram.num_bins))
